Moved The Time Comparison Method
